requirements.txt
selenium
webdriver-manager
python-dotenv
langchain
openai
pyperclip
